requirements.txt
langchain==0.3.0
langchain-core==0.3.0
langchain-huggingface==0.0.48
transformers==4.40.2
torch==2.2.2
sentence-transformers==3.0.1
python-dotenv==1.0.0
pydantic==2.5.3
pytest==7.4.4